requirements.txt
aiogram==3.17.0
torch==2.5.1
torchvision==0.20.1
pillow==11.1.0